closing sidebar on item click
